Add decorator-function for repeating methods in Pokemon class
This decorator @repeat provides the functionality of repeating the execution
of a method a random number of times, if the corresponding method is decorated.
The random number is generated ranging from 0-10. This decorator is used with
speak method in the Pokemon class. A unit test is included for ensuring that the
number of times the speak method is repeated does not exceed maximum number of
repeats which is 10.