Implement changes relating to comments in pull request.
These changes pertains to the comments indicated in the pull request: "Repeat decorator #4".

Added the use of f-strings with print(). Changed the decorator-function to generate a random number from
inside the function, resulting in a new random number being generated each time the speak() method is called.

Contributes to: DGDT-26