Create module for providing client-server communication between pokemon.
This module was created to provide a means for two pokemon objects to communicate with
each other through sockets. This module creates a new Comms class which inherits from
the ElectricPokemon class. Co-routines are added to the class for implenting client-server
communication through sockets and asyncio. A minor change was also made to the pokemon
class in pokemon.py.

Contributes to: DGDT-35