Implement an aiokatcp client and server.
A server class is created which derives from the aiokatcp DeviceServer class. The client is implemented using an object from the aiokatcp.Client class. The server responds accordingly to requests made by the client.

This code demonstrates the use of the aiokatcp module for creating an aiokatcp client and server, and
forms part of the DSP training.

Contributes to: DGDT-48